requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langgraph
groq
faiss-cpu
chromadb
streamlit
python-dotenv
wikipedia
beautifulsoup4
requests
supabase
python-multipart
tiktoken
sentence-transformers
torch
transformers
python-docx
PyPDF2